packages/api/requirements.txt
fastapi>=0.109.0
python-multipart>=0.0.9
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
motor>=3.3.0
redis>=5.0.0
httpx>=0.26.0
feedparser>=6.0.0
structlog>=24.1.0
groq>=0.4.0
qdrant-client>=1.7.0
openai>=1.12.0
bcrypt>=4.1.0
PyJWT>=2.8.0
rapidfuzz>=3.6.0
weasyprint>=60.0
stripe>=8.0.0
apscheduler>=3.10.0
resend>=2.0.0
